+++ src/test/java/cart/cart/controller/CartControllerTest.java
@@ -1,6 +1,7 @@
-package cart.controller;
+package cart.cart.controller;
 
 import cart.auth.AuthSubjectArgumentResolver;
+import cart.cart.controller.CartController;
 import cart.cart.dto.CartProductResponse;
 import cart.cart.dto.CartResponse;
 import cart.cart.service.CartService;
