+++ src/test/java/cart/product/controller/ProductControllerIntegratedTest.java
@@ -1,103 +1,140 @@
-package cart.controller;
+package cart.product.controller;
 
-import cart.config.TestConfig;
-import cart.product.service.ProductMemoryService;
+import cart.config.DBTransactionExecutor;
+import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.util.HashMap;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.then;
-import static org.mockito.Mockito.only;
 
 @SuppressWarnings("NonAsciiCharacters")
-@ContextConfiguration(classes = TestConfig.class)
-@WebMvcTest(ProductController.class)
-class ProductControllerTest {
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class ProductControllerIntegratedTest {
     private static final String DEFAULT_PATH = "/products/";
     
-    @MockBean
-    private ProductMemoryService productMemoryService;
+    @LocalServerPort
+    private int port;
     
-    private HashMap<Object, Object> productRequestMapper;
+    @RegisterExtension
+    private DBTransactionExecutor dbTransactionExecutor;
+    
+    private HashMap<String, Object> productRequestMapper;
+    
+    @Autowired
+    public ProductControllerIntegratedTest(final JdbcTemplate jdbcTemplate) {
+        this.dbTransactionExecutor = new DBTransactionExecutor(jdbcTemplate);
+    }
     
     @BeforeEach
     void setUp() {
-        productRequestMapper = new HashMap<>();
-        
-        final Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-        RestAssuredMockMvc.standaloneSetup(
-                MockMvcBuilders.standaloneSetup(new ProductController(productMemoryService))
-                        .setControllerAdvice(new GlobalExceptionHandler(logger))
-        );
+        productRequestMapper  = new HashMap<>();
+        RestAssured.port = port;
     }
     
     @Test
     void 상품을_생성한다() {
         // given
-        normalInput();
+        productRequestMapper.put("name", "product");
+        productRequestMapper.put("imageUrl", "abel.com");
+        productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .assertThat()
-                .status(HttpStatus.CREATED);
-        
-        then(productMemoryService).should(only()).save(any());
+                .contentType(ContentType.JSON)
+                .statusCode(HttpStatus.CREATED.value());
     }
     
     @Test
     void 상품을_수정한다() {
         // given
-        normalInput();
+        final Integer id = saveAndGetProductId();
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
-                .when().put(DEFAULT_PATH + "1")
+                .when().put(DEFAULT_PATH + id)
                 .then().log().all()
-                .assertThat()
-                .status(HttpStatus.NO_CONTENT);
-        
-        then(productMemoryService).should(only()).update(anyLong(), any());
-    }
-    
-    private void normalInput() {
-        productRequestMapper.put("name", "book");
-        productRequestMapper.put("imageUrl", "abel.com");
-        productRequestMapper.put("price", 10_000);
+                .statusCode(HttpStatus.NO_CONTENT.value());
     }
     
     @Test
     void 상품을_삭제한다() {
-        // expect
-        RestAssuredMockMvc.given().log().all()
-                .when().delete(DEFAULT_PATH + "1")
-                .then().log().all()
-                .assertThat()
-                .status(HttpStatus.NO_CONTENT);
+        // given
+        final Integer id = saveAndGetProductId();
         
-        then(productMemoryService).should(only()).delete(anyLong());
+        // expect
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .when().delete(DEFAULT_PATH + id)
+                .then().log().all()
+                .statusCode(HttpStatus.NO_CONTENT.value());
+    }
+    
+    private Integer saveAndGetProductId() {
+        // given
+        productRequestMapper.put("name", "product");
+        productRequestMapper.put("imageUrl", "abel.com");
+        productRequestMapper.put("price", 1000);
+        
+        // expect
+        return RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequestMapper)
+                .when().post(DEFAULT_PATH)
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value())
+                .contentType(ContentType.JSON)
+                .extract()
+                .response()
+                .path("id");
+    }
+    
+    @Test
+    void 상품을_수정할_시_존재하지_않는_product_id를_전달하면_예외가_발생한다() {
+        // given
+        productRequestMapper.put("name", "product");
+        productRequestMapper.put("imageUrl", "abel.com");
+        productRequestMapper.put("price", 1000);
+        
+        // expect
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequestMapper)
+                .when().put(DEFAULT_PATH + 1)
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 존재하지 않는 product id 입니다."));
+    }
+    
+    @Test
+    void 상품을_삭제할_시_존재하지_않는_product_id를_전달하면_예외가_발생한다() {
+        // expect
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .when().delete(DEFAULT_PATH + 1)
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 존재하지 않는 product id 입니다."));
     }
     
     @ParameterizedTest(name = "{displayName} : name = {0}")
@@ -109,13 +146,13 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
-                .status(HttpStatus.BAD_REQUEST)
                 .body("message", is("[ERROR] 상품 이름을 입력해주세요."));
     }
     
@@ -127,12 +164,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 상품 이름은 255자까지 입력가능합니다."));
     }
@@ -145,12 +182,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 이미지 URL을 입력해주세요."));
     }
@@ -163,34 +200,34 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
-                .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."))
-                .body("message", containsString("[ERROR] 이미지 URL을 입력해주세요."));
+                .body("message", containsString("[ERROR] 이미지 URL을 입력해주세요."))
+                .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."));
     }
     
     @Test
     void 상품_저장_시_이미지_URL_길이가_255초과일때_예외_발생() {
         // given
         productRequestMapper.put("name", "아벨");
-        productRequestMapper.put("imageUrl", "a".repeat(256));
+        productRequestMapper.put("imageUrl", "a".repeat(256) + ".com");
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
-                .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."))
-                .body("message", containsString("[ERROR] 이미지 URL은 255자까지 입력가능합니다."));
+                .body("message", containsString("[ERROR] 이미지 URL은 255자까지 입력가능합니다."))
+                .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."));
     }
     
     @ParameterizedTest(name = "{displayName} : imageUrl = {0}")
@@ -202,13 +239,13 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
                 .contentType(ContentType.JSON)
-                .status(HttpStatus.CREATED);
+                .statusCode(HttpStatus.CREATED.value());
     }
     
     @ParameterizedTest(name = "{displayName} : imageUrl = {0}")
@@ -220,13 +257,13 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
                 .contentType(ContentType.JSON)
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .body("message", is("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."));
     }
     
@@ -238,12 +275,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", null);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격을 입력해주세요."));
     }
@@ -256,12 +293,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", -1);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격의 최소 금액은 0원입니다."));
     }
@@ -274,12 +311,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 10_000_001);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격의 최대 금액은 1000만원입니다."));
     }
@@ -292,12 +329,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", "10000001000");
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().post(DEFAULT_PATH)
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 입력 형식이 잘못되었습니다."));
     }
@@ -311,12 +348,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 상품 이름을 입력해주세요."));
     }
@@ -329,12 +366,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 상품 이름은 255자까지 입력가능합니다."));
     }
@@ -347,12 +384,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 이미지 URL을 입력해주세요."));
     }
@@ -365,12 +402,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", containsString("[ERROR] 이미지 URL을 입력해주세요."))
                 .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."));
@@ -380,16 +417,16 @@ class ProductControllerTest {
     void 상품_수정_시_이미지_URL_길이가_255초과일때_예외_발생() {
         // given
         productRequestMapper.put("name", "아벨");
-        productRequestMapper.put("imageUrl", "a".repeat(256));
+        productRequestMapper.put("imageUrl", "a".repeat(256) + ".com");
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", containsString("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."))
                 .body("message", containsString("[ERROR] 이미지 URL은 255자까지 입력가능합니다."));
@@ -399,18 +436,19 @@ class ProductControllerTest {
     @ValueSource(strings = {"http://abel.com", "https://abel.com", "http://www.abel.com", "https://www.abel.com", "www.abel.com", "abel.com"})
     void 상품_수정_시_이미지_URL_올바른_형식_입력(String imageUrl) {
         // given
+        final Integer productId = saveAndGetProductId();
+        
         productRequestMapper.put("name", "아벨");
         productRequestMapper.put("imageUrl", imageUrl);
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
-                .when().put(DEFAULT_PATH + "1")
+                .when().put(DEFAULT_PATH + productId)
                 .then().log().all()
-                .assertThat()
-                .status(HttpStatus.NO_CONTENT);
+                .statusCode(HttpStatus.NO_CONTENT.value());
     }
     
     @ParameterizedTest(name = "{displayName} : imageUrl = {0}")
@@ -422,13 +460,13 @@ class ProductControllerTest {
         productRequestMapper.put("price", 1000);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
                 .contentType(ContentType.JSON)
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .body("message", is("[ERROR] 이미지 URL의 형식이 올바르지 않습니다."));
     }
     
@@ -440,12 +478,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", null);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격을 입력해주세요."));
     }
@@ -458,12 +496,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", -1);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격의 최소 금액은 0원입니다."));
     }
@@ -476,12 +514,12 @@ class ProductControllerTest {
         productRequestMapper.put("price", 10_000_001);
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 가격의 최대 금액은 1000만원입니다."));
     }
@@ -490,16 +528,16 @@ class ProductControllerTest {
     void 상품_수정_시_price의_자릿수가_Integer_범위를_초과했을_때_예외_발생() {
         // given
         productRequestMapper.put("name", "아벨");
-        productRequestMapper.put("imageUrl", "a");
+        productRequestMapper.put("imageUrl", "abel.com");
         productRequestMapper.put("price", "10000001000");
         
         // expect
-        RestAssuredMockMvc.given().log().all()
+        RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(productRequestMapper)
                 .when().put(DEFAULT_PATH + "1")
                 .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
+                .statusCode(HttpStatus.BAD_REQUEST.value())
                 .contentType(ContentType.JSON)
                 .body("message", is("[ERROR] 입력 형식이 잘못되었습니다."));
     }
